fixed freeze on windows
